Prevent segfault when httpContext is null
When the constructor fails, httpContext is null. But this is not guarded for in

some member functions making a program to segfault when using the builder pattern and
chaining calls to member functions.

This commit adds safeguards so that calls to member functions of a 'failed' object
just returns. This also handles the case when the object has been moved and a call
is made to an member function of the now 'empty' object.